[#163720568] Rename controllers
